# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FSTS Client Dashboard Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sites/{siteId}/courses:
    parameters:
      - name: siteId
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: listCourses
      tags: [content]
      summary: List courses for a site
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [content]
      summary: Create a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /sites/{siteId}/courses/{courseId}:
    parameters:
      - name: siteId
        in: path
        required: true
        schema:
          type: integer
      - name: courseId
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getCourse
      tags: [content]
      summary: Get a single course
      responses:
        "200":
          description: Course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    patch:
      operationId: updateCourse
      tags: [content]
      summary: Update a course (creates a version snapshot)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [content]
      summary: Delete a course (requires confirmation client-side)
      responses:
        "204":
          description: Deleted
...
  /sites/{siteId}/articles:
    parameters:
      - name: siteId
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: listArticles
      tags: [content]
      summary: List articles / blog posts for a site
      responses:
        "200":
          description: List of articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
    post:
      operationId: createArticle
      tags: [content]
      summary: Create an article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
  /sites/{siteId}/articles/{articleId}:
    parameters:
      - name: siteId
        in: path
        required: true
        schema:
          type: integer
      - name: articleId
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getArticle
      tags: [content]
      summary: Get a single article
      responses:
        "200":
          description: Article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
    patch:
      operationId: updateArticle
      tags: [content]
      summary: Update an article (creates a version snapshot)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
    delete:
      operationId: deleteArticle
      tags: [content]
      summary: Delete an article (requires confirmation client-side)
      responses:
        "204":
          description: Deleted
...
  /sites/{siteId}/email-settings:
    parameters:
      - name: siteId
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getEmailSettings
      tags: [email]
      summary: Get email settings for a site
      responses:
        "200":
          description: Email settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailSettings"
    put:
      operationId: updateEmailSettings
      tags: [email]
      summary: Update email settings for a site
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailSettingsInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailSettings"
...
  /sites/{siteId}/backups:
    parameters:
      - name: siteId
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: listBackups
      tags: [backups]
      summary: List backups for a site
      responses:
        "200":
          description: List of backups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Backup"
    post:
      operationId: createBackup
      tags: [backups]
      summary: Create a manual backup snapshot of a site's content
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Backup"
  /sites/{siteId}/backups/{backupId}/restore:
    parameters:
      - name: siteId
        in: path
        required: true
        schema:
          type: integer
      - name: backupId
        in: path
        required: true
        schema:
          type: integer
    post:
      operationId: restoreBackup
      tags: [backups]
      summary: Restore a site from a backup snapshot (requires confirmation client-side)
      responses:
        "200":
          description: Restored backup
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Backup"
components:
  schemas:
...
    Course:
      type: object
      required: [id, siteId, title, slug, status, description, createdAt, updatedAt]
      properties:
        id: { type: integer }
        siteId: { type: integer }
        title: { type: string }
        slug: { type: string }
        status:
          type: string
          enum: [draft, published, archived]
        description: { type: string }
        durationLabel: { type: ["string", "null"] }
        priceCents: { type: ["integer", "null"] }
        imageUrl: { type: ["string", "null"] }
        squareItemId: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CourseInput:
      type: object
      required: [title, slug, description]
      properties:
        title: { type: string, minLength: 1 }
        slug: { type: string, minLength: 1 }
        status:
          type: string
          enum: [draft, published, archived]
        description: { type: string, minLength: 1 }
        durationLabel: { type: string }
        priceCents: { type: integer }
        imageUrl: { type: string }
        squareItemId: { type: string }
    CourseUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        slug: { type: string, minLength: 1 }
        status:
          type: string
          enum: [draft, published, archived]
        description: { type: string, minLength: 1 }
        durationLabel: { type: ["string", "null"] }
        priceCents: { type: ["integer", "null"] }
        imageUrl: { type: ["string", "null"] }
        squareItemId: { type: ["string", "null"] }
...
    Article:
      type: object
      required: [id, siteId, title, slug, status, body, createdAt, updatedAt]
      properties:
        id: { type: integer }
        siteId: { type: integer }
        title: { type: string }
        slug: { type: string }
        status:
          type: string
          enum: [draft, published, archived]
        excerpt: { type: ["string", "null"] }
        body: { type: string }
        coverImageUrl: { type: ["string", "null"] }
        publishedAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ArticleInput:
      type: object
      required: [title, slug, body]
      properties:
        title: { type: string, minLength: 1 }
        slug: { type: string, minLength: 1 }
        status:
          type: string
          enum: [draft, published, archived]
        excerpt: { type: string }
        body: { type: string, minLength: 1 }
        coverImageUrl: { type: string }
        publishedAt: { type: string, format: date-time }
    ArticleUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        slug: { type: string, minLength: 1 }
        status:
          type: string
          enum: [draft, published, archived]
        excerpt: { type: ["string", "null"] }
        body: { type: string, minLength: 1 }
        coverImageUrl: { type: ["string", "null"] }
        publishedAt: { type: ["string", "null"], format: date-time }
...
    EmailSettings:
      type: object
      required: [siteId, fromName, fromEmail, replyToEmail, updatedAt]
      properties:
        siteId: { type: integer }
        fromName: { type: string }
        fromEmail: { type: string }
        replyToEmail: { type: string }
        notifyOnNewLead: { type: boolean }
        notifyOnBooking: { type: boolean }
        updatedAt: { type: string, format: date-time }
    EmailSettingsInput:
      type: object
      required: [fromName, fromEmail, replyToEmail]
      properties:
        fromName: { type: string, minLength: 1 }
        fromEmail: { type: string, minLength: 1 }
        replyToEmail: { type: string, minLength: 1 }
        notifyOnNewLead: { type: boolean }
        notifyOnBooking: { type: boolean }
...
    Backup:
      type: object
      required: [id, siteId, label, sizeBytes, createdAt]
      properties:
        id: { type: integer }
        siteId: { type: integer }
        label: { type: string }
        sizeBytes: { type: integer }
        createdAt: { type: string, format: date-time }